# swagger.yaml
swagger: '2.0'
info: 
  version: 1.0.0
  title: My Cool Node Swagger Express Example
basePath: /v1
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - https
paths:
  '/swagger.json':
    get:
      tags:
        - Spec
      operationId: getSwaggerSpec
      description: | 
        Get Swagger Spec
      responses:
        '200':
          description: Successful
          schema:
            type: string
  '/register':
    post:
      tags:
        - Register
      operationId: Register
      parameters:
        - name: user
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
              address:
                type: string
              password:
                type: string
      description: | 
        Register User
      responses:
        '200':
          description: Successful
          schema:
            type: string
    get:
      tags:
        - Register
      operationId: GetRegister
      description: | 
        Get Register User
      responses:
        '200':
          description: Successful
          schema:
            type: string
  '/register/{id}':
    get:
      tags:
        - Register
      operationId: GetRegisteredUserByID
      parameters:
        - $ref: '#/parameters/id'
      description: | 
        Get Register User
      responses:
        '200':
          description: Successful
          schema:
            type: string
    put:
      tags:
        - Register
      operationId: UpdateRegisteredUserByID
      parameters:
        - $ref: '#/parameters/id'
        - name: user
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
              address:
                type: string
              password:
                type: string
      description: | 
        Get Register User
      responses:
        '200':
          description: Successful
          schema:
            type: string
    delete:
      tags:
        - Register
      operationId: DeleteRegisteredUserByID
      parameters:
        - $ref: '#/parameters/id'
      description: | 
        Delete Register User
      responses:
        '200':
          description: Successful
          schema:
            type: string            
parameters:
  id:
    in: path
    name: id
    required: true
    type: string  
  